KVM: arm64: Introduce support to pin VMIDs
Introduce kvm_arm_pinned_vmid_get() and kvm_arm_pinned_vmid_put(), to pin
a VMID associated with a KVM instance. This will guarantee that VMID
remains same after a rollover.

On an ARM64 system with a SMMUv3 implementation that fully supports
Broadcast TLB Maintenance(BTM) feature, the CPU TLB invalidate
instructions are received by SMMUv3. This is very useful when the
SMMUv3 shares the page tables with the CPU(eg: Guest SVA use case).
But for this to work SMMUv3 should use the VMID used by KVM for
its s2 stage.

These new interfaces provides a way for SMMUv3 to retrieve the KVM VMID
for its s2 stage and guarantees that it remain same even after VMID
rollovers.

Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>