iommufd: Add IOMMU_HWPT_TYPE_KERNEL type
This adds IOMMU_HWPT_TYPE_KERNEL for userspace to allocate kernel-managed
domain with a specified IOPTE format.

On the platforms that support nested translation, there are two formats of
IOPTEs, stage-1 and stage-2. Userspace could use either of them for single
stage translation usages. This adds the IOMMU_HWPT_TYPE_KERNEL type to
accept user data for the hwpt allocation, and defines a flag to tell kernel
which IOPTE format is wanted for the domain.

Signed-off-by: Yi Liu <yi.l.liu@intel.com>