iommufd: Interface to associate kvm* with iommufd_ctx
The interface takes a struct kvm* and assigns it to struct iommufd_ctx
kvm* safely by acquiring a kvm ref. It releases the kvm reference when
the iommufd_ctx gets freed.

This is in preparation to provide an interface for vfio_device to set
the corresponding kvm* if it uses the iommufd backend. Future changes
to iommufd will add support to pass the kvm* to iommu drivers for use
in domain configuration/setup.

Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>