vfio: Add struct to hold KVM assets and dedup group vs. iommufd code
Add a struct to hold the KVM assets need to manage and pass along KVM
references to VFIO devices.  Providing a common struct deduplicates the
group vs. iommufd code, and will make it easier to rework the attachment
logic so that VFIO doesn't have to do a symbol lookup to retrieve the
get/put helpers from KVM.

Reviewed-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Sean Christopherson <seanjc@google.com>
Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>